MODE-1666 - Changed back the name of JndiRepositoryFactory for the real JNDI factory
